Code sur ctrlComment. +++

+++ Serveur/controllers/ctrlComment.js
@@ -11,9 +11,7 @@ exports.getAllcommentsUser = async (req, res) => {
       res.json(comments);
    } catch (error) {
       res.status(500).send({
-         message:
-            error.message ||
-            'Une erreur est survenue dans la recherche de comments.',
+         message: error.message || 'Une erreur est survenue dans la recherche de comments.',
       });
    }
 };
@@ -29,61 +27,111 @@ exports.getOnecommentUser = async (req, res) => {
       res.json(comment);
    } catch (error) {
       res.status(500).send({
-         message:
-            error.message ||
-            'Une erreur est survenue dans la recherche de comment.',
+         message: error.message || 'Une erreur est survenue dans la recherche de comment.',
       });
    }
 };
 
 exports.createcommentUser = async (req, res) => {
+   const { body, postId } = req.body;
+   const userId = req.session.user.id;
+
+   if (!body || body == '' || !postId) {
+      return res.status(400).send('Le commentaire ne peut être vide');
+   }
+
    try {
       const comment = await prisma.comments.create({
-         data: req.body,
+         data: {
+            body: body,
+            postId: Number(postId),
+            userId: Number(userId),
+         },
       });
-      res.json(comment);
+      res.status(201).json(comment);
    } catch (error) {
       res.status(500).send({
-         message:
-            error.message ||
-            'Une erreur est survenue dans la création de comment.',
+         message: error.message || 'Une erreur est survenue dans la création de comment.',
       });
    }
 };
 
 exports.updatecommentUser = async (req, res) => {
-   const { id } = req.params;
+   const commentId = req.params.id;
+   const { body, postId } = req.body;
+
+   if (!body || body == '' || !commentId) {
+      return res.status(400).send('Le commentaire ne peut être vide');
+   }
+
    try {
+      // Récupération du propriétaire du commentaire sur la Bd
+      const dataCommentBd = await findOneComment(commentId);
+
+      if (!dataCommentBd) {
+         return res.status(404).send('Commentaire non trouvé');
+      }
+
+      if (parseInt(req.session.user.id) !== dataCommentBd.userId) {
+         return res.status(401).send('Erreur de user');
+      }
+
+      if (postId !== dataCommentBd.postId) {
+         return res.status(401).send('Erreur de post');
+      }
+
       const comment = await prisma.comments.update({
          where: {
-            id: Number(id),
+            id: Number(commentId),
          },
-         data: req.body,
+         data: {
+            body: body,
+         },
       });
-      res.json(comment);
+      res.status(201).json(comment);
    } catch (error) {
       res.status(500).send({
-         message:
-            error.message ||
-            'Une erreur est survenue dans la modification de comment.',
+         message: error.message || 'Une erreur est survenue dans la modification de comment.',
       });
    }
 };
 
 exports.deletecommentUser = async (req, res) => {
+   const commentId = req.params.id;
    try {
-      const { id } = req.params;
+      // Récupération du propriétaire du commentaire sur la Bd
+      const dataCommentBd = await findOneComment(commentId);
+
+      if (!dataCommentBd) {
+         return res.status(404).send('Commentaire non trouvé');
+      }
+
+      // Vérification du propriétaire du commentaire
+      if (parseInt(req.session.user.id) !== dataCommentBd.userId) {
+         return res.status(401).send('Erreur de user');
+      }
+
       const comment = await prisma.comments.delete({
          where: {
-            id: Number(id),
+            id: Number(commentId),
          },
       });
-      res.json(comment);
+      res.status(200).send('Commentaire supprimé');
    } catch (error) {
       res.status(500).send({
-         message:
-            error.message ||
-            'Une erreur est survenue dans la suppression de comment.',
+         message: error.message || 'Une erreur est survenue dans la suppression de comment.',
       });
    }
 };
+
+async function findOneComment(commentId) {
+   return await prisma.comments.findUnique({
+      where: {
+         id: Number(commentId),
+      },
+      select: {
+         userId: true,
+         postId: true,
+      },
+   });
+}
